Silence ICC, GCC/SLC5, clang/MacOS warnings
git-svn-id: http://root.cern.ch/svn/root/trunk@39411 27541ba8-7e3a-0410-8455-c3a389f83636

+++ core/textinput/src/textinput/Editor.cpp
@@ -98,7 +98,7 @@ namespace textinput {
       return true;
     }
 
-    fCurHistEntry = -1;
+    fCurHistEntry = (size_t) -1;
     return false;
   }
 
@@ -205,7 +205,7 @@ namespace textinput {
         return kPRSuccess;
       case kCmdEnter:
         fReplayHistEntry = (size_t) -1;
-        fCurHistEntry = -1;
+        fCurHistEntry = (size_t) -1;
         CancelSpecialInputMode(R.fDisplay);
         return kPRSuccess;
       case kCmdDelLeft:
